+++ honglidong/app/src/main/java/com/hongildong/map/navGraph/SearchNavHost.kt
@@ -3,8 +3,6 @@ package com.hongildong.map.navGraph
 import android.util.Log
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
-import androidx.core.graphics.component1
-import androidx.core.graphics.component2
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavHostController
 import androidx.navigation.NavType
@@ -12,7 +10,6 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.navArgument
 import androidx.navigation.navigation
-import com.hongildong.map.data.entity.ReviewInfo
 import com.hongildong.map.data.entity.SearchKeyword
 import com.hongildong.map.data.entity.SearchableNodeType
 import com.hongildong.map.ui.bookmark.BookmarkViewModel
@@ -23,8 +20,8 @@ import com.hongildong.map.ui.search.direction.DirectionScreen
 import com.hongildong.map.ui.search.direction.DirectionSearchScreen
 import com.hongildong.map.ui.search.location_detail.facility.FacilityDetailScreen
 import com.hongildong.map.ui.search.location_detail.LocationDetailScreen
-import com.hongildong.map.ui.search.location_detail.facility.ReviewScreen
-import com.hongildong.map.ui.search.location_detail.facility.ReviewViewModel
+import com.hongildong.map.ui.search.location_detail.facility.review.ReviewScreen
+import com.hongildong.map.ui.search.location_detail.facility.review.ReviewViewModel
 import com.hongildong.map.ui.util.bottomsheet.BottomSheetViewModel
 import com.hongildong.map.ui.util.map.MapViewmodel
